created models and migration for equipments
