+++ collectives/templates/equipment.html
@@ -45,8 +45,8 @@
 </script>
 {% endif %}
 
-{% for user in users %}
-  {{  user }}
-{% endfor %}
+
+{{  equipment.reference }}
+
 
 {% endblock %}
